requirement.txt
fastapi
rasa
textblob
spacytextblob
spacy
httpx
pyyaml
typing
uvicorn
nltk
python-dotenv
urllib3
glob2
python-multipart